requirements.txt
streamlit
langchain
langchain-openai
langgraph
fpdf2
python-dotenv
pydantic
